Move common functionality of koji class to KojiBase
